requirements.txt
# Traffic Red Light Running Violation Detection System Requirements
# Usage: pip install -r requirements.txt

# Base ----------------------------------------
matplotlib>=3.2.2
opencv-python>=4.7.0
Pillow>=8.0.0
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
torch>=1.13.0
torchvision>=0.14.0
tqdm>=4.64.0
numpy>=1.21.0

# Computer Vision and AI ----------------------
ultralytics>=8.0.0
cvzone
deep-sort-realtime

# OCR and Text Detection ----------------------
easyocr>=1.7.0
paddleocr>=2.6.0

# Flask and Web Framework ---------------------
Flask>=2.3.3
flask-cors>=4.0.0
Flask-MySQLdb>=1.0.1
werkzeug>=2.3.7

# Database ------------------------------------
mysqlclient>=2.2.0
PyMySQL>=1.0.0

# Data Analysis and Plotting ------------------
pandas>=1.1.4
seaborn>=0.11.0

# Image Processing ----------------------------
scikit-image>=0.19.0

# PDF Generation ------------------------------
reportlab>=4.0.4

# System Utilities ----------------------------
psutil

# Async Processing ----------------------------
aiofiles>=0.8.0

# Development and Testing ---------------------
pytest>=7.0.0
pytest-asyncio>=0.20.0

# Logging -------------------------------------
# tensorboard>=2.13.0
# clearml
# comet

# Export --------------------------------------
# coremltools>=6.0  # CoreML export
# onnx>=1.12.0  # ONNX export
# onnxsim>=0.4.1  # ONNX simplifier
# nvidia-pyindex  # TensorRT export
# nvidia-tensorrt  # TensorRT export
# scikit-learn==0.19.2  # CoreML quantization
# tensorflow>=2.4.1  # TF exports (-cpu, -aarch64, -macos)
# tflite-support
# tensorflowjs>=3.9.0  # TF.js export
# openvino-dev>=2022.3  # OpenVINO export

# Additional Extras ---------------------------
# thop>=0.1.1  # FLOPs computation
# ipython  # interactive notebook
# albumentations>=1.0.3
# pycocotools>=2.0.6  # COCO mAP
# roboflow